KBV-101: Move question and answer lambda
- the lambdas in the handler folder in the lambdas folder
