+++ lambdas/questions/build.gradle
@@ -1,33 +1,54 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java library project to get you started.
- * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
- * User Manual available at https://docs.gradle.org/7.3.3/userguide/building_java_projects.html
- */
-
 plugins {
-    // Apply the java-library plugin for API and implementation separation.
-    id 'java-library'
+	id "java"
+	id "application"
+	id "idea"
+	id "jacoco"
+	id "io.freefair.aspectj.post-compile-weaving" version "6.3.0"
 }
 
 repositories {
-    // Use Maven Central for resolving dependencies.
-    mavenCentral()
+	mavenCentral()
 }
 
 dependencies {
-    // Use JUnit Jupiter for testing.
-    testImplementation 'org.junit.jupiter:junit-jupiter:5.7.2'
+	implementation "com.amazonaws:aws-lambda-java-core:1.2.1",
+			"com.amazonaws:aws-lambda-java-events:3.11.0",
+			'com.fasterxml.jackson.core:jackson-databind:2.13.1',
+			'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.13.1',
+			'org.apache.cxf:cxf-rt-frontend-jaxws:3.4.5',
+			'org.apache.cxf:cxf-rt-transports-http:3.4.5',
+			'javax.annotation:javax.annotation-api:1.3.2',
+			'org.apache.logging.log4j:log4j-api:2.17.1',
+			'org.apache.logging.log4j:log4j-core:2.17.1',
+			'com.sun.xml.messaging.saaj:saaj-impl:1.5.3',
+			"org.hibernate.validator:hibernate-validator:6.2.1.Final",
+			"org.hibernate:hibernate-validator-annotation-processor:6.2.1.Final",
+			"javax.validation:validation-api:2.0.1.Final",
+			"javax.el:javax.el-api:3.0.0",
+			"org.glassfish:javax.el:3.0.0",
+			project(":lib")
 
-    // This dependency is exported to consumers, that is to say found on their compile classpath.
-    api 'org.apache.commons:commons-math3:3.6.1'
 
-    // This dependency is used internally, and not exposed to consumers on their own compile classpath.
-    implementation 'com.google.guava:guava:30.1.1-jre'
+
+	aspect "software.amazon.lambda:powertools-tracing:1.10.2",
+			"software.amazon.lambda:powertools-parameters:1.10.2"
+
+	runtimeOnly 'org.apache.logging.log4j:log4j-slf4j18-impl:2.17.1',
+			'com.amazonaws:aws-lambda-java-log4j2:1.5.1'
+
+	testImplementation "org.junit.jupiter:junit-jupiter:5.8.2",
+			'org.mockito:mockito-junit-jupiter:4.3.1',
+			'org.mockito:mockito-inline:4.3.1'
 }
 
-tasks.named('test') {
-    // Use JUnit Platform for unit tests.
-    useJUnitPlatform()
+test {
+	useJUnitPlatform ()
+	finalizedBy jacocoTestReport
+}
+
+jacocoTestReport {
+	dependsOn test
+	reports {
+		xml.required.set(true)
+	}
 }
