--- lambdas/questions/src/main/java/uk/gov/di/ipv/cri/experian/kbv/api/Library.java
@@ -1,10 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package uk.gov.di.ipv.experian.kbv.api;
-
-public class Library {
-    public boolean someLibraryMethod() {
-        return true;
-    }
-}
