--- lambdas/questions/src/test/java/uk/gov/di/ipv/cri/experian/kbv/api/LibraryTest.java
@@ -1,16 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package uk.gov.di.ipv.cri.experian.kbv.api;
-
-import org.junit.jupiter.api.Test;
-import uk.gov.di.ipv.experian.kbv.api.Library;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class LibraryTest {
-    @Test void someLibraryMethodReturnsTrue() {
-        Library classUnderTest = new Library();
-        assertTrue(classUnderTest.someLibraryMethod(), "someLibraryMethod should return 'true'");
-    }
-}
