+++ lib/src/main/resources/wsdl/iiq-service.wsdl
