+++ lib/src/main/resources/wsdl/iiq-wasp-token.wsdl
