+++ settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = "di-ipv-experian-kbv-api"
 include "lib",
-        "lambdas",
-        "lambdas:questions",
-        "lambdas:answers"
+		"lambdas",
+		"lambdas:questions",
+		"lambdas:answers"
